make non-windows build also depend on an out-of-source path
This makes pyhanabi.py look for a folder named 'build/' in the current
directory. This is pretty fragile, because it could be in some other
location, and also it is not a canonical way to provide libraries
(usually they are in a system path, or the libraries (but not other
miscellaneous build files) are installed to the working directory). This
could be changed by adding an installation command in the future.